src/websockets/requirements.txt
psycopg2-binary==2.9.2
websockets==10.1
polygon==0.9.7
datadog-api-client